+++ src/web/admin_urls.py
@@ -20,7 +20,7 @@ urlpatterns = [
         staff_member_required,
         'django.conf.urls.i18n'
     )),
-    *static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT),  # for development only; Nginx is used in production
+    *static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT),  # For development only; Nginx is used in production
 
     *ckeditor_uploader_urls(),
 ]